add scopes

+++ app/Http/Controllers/PostController.php
@@ -6,6 +6,7 @@ use App\Models\Post;
 use Illuminate\Support\Str;
 use Illuminate\Http\Request;
 use App\Http\Requests\StorePost;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Facades\DB;
 
 class PostController extends Controller
@@ -24,7 +25,10 @@ class PostController extends Controller
     {
         /* DB::connection()->enableQueryLog();
          dd(DB::getQueryLOg());*/
-        $posts = Post::withCount('comments')->orderBy('updated_at','desc')->get();
+         /*
+            $posts = [Post::withCount('comments')]== bulder->orderBy('updated_at','desc')->get();
+         */
+        $posts = Post::withCount('comments')->get();
         $tab='list';
         return view('posts.index',compact('posts','tab'));
     }
@@ -34,7 +38,7 @@ class PostController extends Controller
     {
         /* DB::connection()->enableQueryLog();
          dd(DB::getQueryLOg());*/
-        $posts = Post::onlyTrashed('comments')->orderBy('updated_at','desc')->get();
+        $posts = Post::onlyTrashed('comments')->get();
         $tab='archive';
         return view(
             'posts.index'
@@ -45,7 +49,7 @@ class PostController extends Controller
     {
         /* DB::connection()->enableQueryLog();
          dd(DB::getQueryLOg());*/
-        $posts = Post::withTrashed()->withCount('comments')->orderBy('updated_at','desc')->get();
+        $posts = Post::withTrashed()->withCount('comments')->get();
         $tab='all';
         return view('posts.index',compact('posts','tab'));
     }
@@ -56,6 +60,7 @@ class PostController extends Controller
      */
     public function create()
     {
+        // $this->authorize('create');
          return view('posts.create');
     }
 
@@ -73,10 +78,11 @@ class PostController extends Controller
              'title'   =>'bail|min:4|required|max:100',
              'content' =>'required'
          ]);*/
-       $data=$request->only(['title','content']);
-       $data['slug']=Str::slug($data['title'],'-');
-       $data['active']=true;
-       $post =Post::create($data);
+          $data=$request->validated();
+          $data['user_id'] = $request->user()->id;
+          //when you want  use create that you have in elequent you have to declare  the the atributes in the model */
+          $post =Post::create($data);
+
 
       $request->session()->flash('status', $post->title.'post was created');
       return redirect()->route('posts.index') ;
@@ -105,6 +111,7 @@ class PostController extends Controller
     public function edit($id)
     {
         $post=Post::findOrFail($id);
+        $this->authorize('update',$post);
        return view("posts.edit",['post'=>$post]);
     }
 
@@ -118,6 +125,7 @@ class PostController extends Controller
     public function update(StorePost $request, $id)
     {
         $post=Post::find($id);
+        $this->authorize("update",$post);
         $post->title   = $request->input('title');
         $post->content = $request->input('content');
         $post->save();
@@ -136,22 +144,23 @@ class PostController extends Controller
     {
 
         $post =Post::find($id);
+        $this->authorize("delete",$post);
         $post->delete();
-
         $request->session()->flash('status','was deleted seccefully');
         return redirect()->route('posts.index');
 
     }
     public function restore($id){
-
-       $post=Post::onlyTrashed()->where('id',$id)->first();
-       $post->restore();
-
-       return back();
+        $post=Post::onlyTrashed()->where('id',$id)->first();
+        $this->authorize('restore', $post);
+        $post->restore();
+        return back();
     }
 
     public function forcedelete($id){
+
        $post=Post::onlyTrashed()->where('id',$id)->first();
+       $this->authorize('forceDelete' , $post);
        $post->forceDelete();
        return back() ;
     }
